Cleanup a lot of code. Globally add prefix component.
