+++ src/Datadog.php
@@ -4,6 +4,9 @@ namespace ChaseConey\LaravelDatadogHelper;
 
 use Illuminate\Support\Facades\Facade;
 
+/**
+ * @see \ChaseConey\LaravelDatadogHelper
+ */
 class Datadog extends Facade
 {
     protected static function getFacadeAccessor()
