+++ src/LaravelDatadogHelperServiceProvider.php
@@ -14,11 +14,13 @@ class LaravelDatadogHelperServiceProvider extends ServiceProvider
      */
     public function boot()
     {
+
         $this->publishes(
             [
                 __DIR__ . '/../config/datadog-helper.php' => config_path('datadog-helper.php'),
             ]
         );
+
     }
 
     /**
@@ -41,6 +43,6 @@ class LaravelDatadogHelperServiceProvider extends ServiceProvider
             config('datadog-helper.statsd_port')
         );
 
-        $this->app->singleton('datadog', \Datadogstatsd::class);
+        $this->app->singleton('datadog', LaravelDatadogHelper::class);
     }
 }
